+++ src/meico/supplementary/RandomNumberProvider.java
@@ -162,7 +162,7 @@ public class RandomNumberProvider {
      */
     public void setSeed(long seed) {
         this.random.setSeed(seed);
-        this.series.clear();
+        this.setInitialValue(this.series.get(this.series.size() - 1));
     }
 
     /**